backend/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
asyncpg==0.30.0
psycopg[binary]==3.2.9
sqlalchemy==2.0.41
greenlet==3.0.3
langchain==0.3.26
langchain-core==0.3.68
langchain-ollama==0.3.4
langchain-postgres==0.0.15
langchain-huggingface==0.3.0
langchain-text-splitters==0.3.8
sentence-transformers==5.0.0
